requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-rest-auth==1.1.0
Django==3.1.13
django-allauth==0.42.0
djangorestframework==3.11.2
drf-yasg==1.17.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
